CLN: Start consolidating cik_lookup
