+++ docs/source/ciklookup.rst
@@ -24,5 +24,5 @@ Accessing ``lookups.lookup_dict`` would then return
 
 Another alternative to using the ``CIKLookup`` class directly is to use the provided :meth:`secedgar.utils.get_cik_map` function.
 
-.. autoclass:: secedgar.filings.cik_lookup.CIKLookup
+.. autoclass:: secedgar.utils.cik_lookup.CIKLookup
    :members:
